packages/backend/pyproject.toml
[project]
name = "arqsoft-lab-backend"
version = "0.1.0"
description = ""
authors = [{ name = "JesuFrancesco" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
